+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/consts/Locales.kt
@@ -1,31 +1,27 @@
 package com.windea.breezeframework.core.consts
 
-/**
- * 语言环境。
- *
- * 为了能够自定义语言环境，不应当使用枚举表示。
- */
+/**语言环境的常量集。*/
 object Locales {
-	const val English: String = "en"
-	const val French: String = "fr"
-	const val German: String = "de"
-	const val Italian: String = "it"
-	const val Japanese: String = "ja"
-	const val Korean: String = "ko"
-	const val Chinese: String = "zh"
-	const val SimpleChinese: String = "zh_CN"
-	const val TraditionalChinese: String = "zh_TW"
+	const val english: String = "en"
+	const val french: String = "fr"
+	const val german: String = "de"
+	const val italian: String = "it"
+	const val japanese: String = "ja"
+	const val korean: String = "ko"
+	const val chinese: String = "zh"
+	const val simpleChinese: String = "zh_CN"
+	const val traditionalChinese: String = "zh_TW"
 	
-	const val France: String = "fr_FR"
-	const val Germany: String = "de_DE"
-	const val Italy: String = "it_IT"
-	const val Japan: String = "ja_JP"
-	const val Korea: String = "ko_KR"
-	const val China: String = "zh_CN"
-	const val PRC: String = "zh_CN"
-	const val Taiwan: String = "zh_TW"
-	const val UK: String = "en_GB"
-	const val US: String = "en_US"
-	const val Canada: String = "en_CA"
-	const val CanadaFrench: String = "fr_CA"
+	const val france: String = "fr_FR"
+	const val germany: String = "de_DE"
+	const val italy: String = "it_IT"
+	const val japan: String = "ja_JP"
+	const val korea: String = "ko_KR"
+	const val china: String = "zh_CN"
+	const val prc: String = "zh_CN"
+	const val taiwan: String = "zh_TW"
+	const val uk: String = "en_GB"
+	const val us: String = "en_US"
+	const val canada: String = "en_CA"
+	const val canadaFrench: String = "fr_CA"
 }