+++ CHANGELOG.md
@@ -117,9 +117,9 @@
 * [X] breeze-core 添加`checkChance`扩展。
 * [X] breeze-dsl 完善MarkdownDsl。添加特性支持。重新实现内联dsl。
 * [X] breeze-core 添加`toStringOrEmpty`扩展。
+* [X] breeze-core 添加`Colors`，提供标准颜色常量集。
 * [ ] breeze-dsl 编写一种泛用而灵活的富文本Dsl的规范。不提供实现。
 * [ ] breeze-dsl 编写一种泛用而灵活的思维导图Dsl的规范。不提供实现。
-* [ ] breeze-dsl 区分DslElement和DslElement，只允许通过字符串模版插入DslElement。
 * [ ] breeze-game 初步完成有限状态机的规范。
 * [ ] breeze-game 初步完成实体&组件的规范。
 * [ ] 提供一种“全局变量池”的实现，类似依赖注入，避免`var foo = ...`写法。
@@ -127,7 +127,6 @@
 * [ ] breeze-core Uri构建和编码的扩展。
 * [ ] breeze-core 利用反射的类型转换的扩展。
 * [ ] breeze-text 英文单数与复数的转化。
-* [ ] breeze-emmx 对于MindMaster的增强，例如json/yaml与emmx文档的相互转化。
 
 # 长期